+++ packages/drizzle/CHANGELOG.md
@@ -2,9 +2,16 @@
 
 All notable changes to this project will be documented in this file.
 
-## next (YYYY-MM-DD)
+## Next (YYYY-MM-DD)
 
 - Refactor: Migrating to Relational Queries version 2
+
+## 0.9.2 (2025-05-13)
+
+- Fix: DrizzleInputFactory to support custom column types
+
+## 0.9.0 (2025-05-11)
+
 - Feat: add `DrizzleResolverFactory.queriesResolver` to create a read-only resolver
 - Feat: add `getSelectedColumns` to get the selected columns from the resolver payload